added is_pending field to orders
added is_pending boolean field to orders. so that orders which are already shipped can be differentiated.

changed the way in which addresses and products are deleted. they won't be deleted from now on, due to the foreign key present.. They will just be Hidden, but present in the DB.

Changed the Residential Delivery to False in FedEx's part due to API Team's requirements.